fix: Lazy load _datashader_bundling into own file
